+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Multi-style plotting capabilities for generated example plots
 - JSON file for plotting themes
 - Smoke testing in relevant tox environments
+- `ContinuousParameter` base class
 
 ### Changed
 - `Recommender`s now share their core logic via their base class